+++ src/main/java/JEstebanC/FastFoodApp/dto/ReportSalesWeeklyDTO.java
@@ -2,11 +2,13 @@ package JEstebanC.FastFoodApp.dto;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 /**
  * @author Juan Esteban Castaño Holguin castanoesteban9@gmail.com 6/13/2022
  */
 @Data
 public class ReportSalesWeeklyDTO {
-    private Double weekday;
-    private int total;
+    private Long weekday;
+    private BigInteger total;
 }
